+++ service/generator/interface.go
@@ -4,7 +4,7 @@ import "geoindexing_comparison/service/geo"
 
 type Generator interface {
 	Point(input *Input) geo.Point
-	Points(input *Input, amount int) geo.Points
+	Points(input *Input, amount uint64) geo.Points
 }
 
 func All() []Generator {
