+++ service/index/kdtree/addapter.go
@@ -72,9 +72,9 @@ func (r *CollectionKDTree) RangeSearchTimed(
 	return toConcrete(res), dur
 }
 
-func (r *CollectionKDTree) KNNTimed(point geo.Point, n int) (geo.Points, time.Duration) {
+func (r *CollectionKDTree) KNNTimed(point geo.Point, n uint64) (geo.Points, time.Duration) {
 	t0 := time.Now()
-	res := r.impl.KNN(&point, n)
+	res := r.impl.KNN(&point, int(n))
 	dur := time.Now().Sub(t0)
 
 	return toConcrete(res), dur
